+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/settings/ZigConfigurable.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of ZigBrains.
  *
- * Copyright (C) 2023-2024 FalsePattern
+ * Copyright (C) 2023-2025 FalsePattern
  * All Rights Reserved
  *
  * The above copyright notice and this permission notice shall be included
@@ -22,11 +22,10 @@
 
 package com.falsepattern.zigbrains.project.settings
 
-import com.falsepattern.zigbrains.lsp.settings.ZLSSettingsConfigurable
 import com.falsepattern.zigbrains.shared.MultiConfigurable
 import com.intellij.openapi.project.Project
 
-class ZigConfigurable(project: Project): MultiConfigurable(ZigProjectConfigurable(project), ZLSSettingsConfigurable(project)) {
+class ZigConfigurable(project: Project): MultiConfigurable(ZigProjectConfigurationProvider.createConfigurables(project)) {
     override fun getDisplayName(): String {
         return "Zig"
     }